Add duration events and create event card.
Create component event card.
Add attribute to constructor event card.
Replace event html with event component.
Create Evento Factory.
Modify DatabaseSeeder to run Evento Factory.
Add duration of the event to migration.
Add hour and date methods in Evento model.
Add duracion_hours to factory evento.
Add style to evento component.
Add validation to EventoRequest.
Add logic only to be delete or edit it by owner.
Add field duracion_horas to create and edit form.

+++ app/Http/Controllers/EventoController.php
@@ -67,8 +67,11 @@ class EventoController extends Controller
      */
     public function update(EventoRequest $request, Evento $evento)
     {
+        if( Auth::id() != $evento->user_id){
+            return redirect()->route('eventos.index');
+        }
         $evento->update($request->all());
-        return redirect()->route('comunidades.show',$evento->comunidad->user_id);
+        return redirect()->route('eventos.index');
     }
 
     /**
@@ -76,25 +79,28 @@ class EventoController extends Controller
      */
     public function destroy(Evento $evento)
     {
+        if( Auth::id() != $evento->user_id){
+         return redirect()->route('eventos.index');
+        }
         $evento->delete();
-        return redirect()->route('comunidades.show',$evento->comunidad->user_id);
+        return redirect()->route('eventos.index');
     }
 
     public function aceptar(Evento $evento)
     {
         if($evento->comunidad->user_id != Auth::id()){
-            return redirect()->route('comunidades.show',$evento->comunidad->user_id);
+            return redirect()->route('eventos.index');
         }
         $evento->update(["estado_moderacion" => "aprobado"]);
-        return redirect()->route('comunidades.show',$evento->comunidad->user_id);
+        return redirect()->route('eventos.index');
     }
 
     public function rechazar(Evento $evento)
     {
         if($evento->comunidad->user_id != Auth::id()){
-            return redirect()->route('comunidades.show',$evento->comunidad->user_id);
+            return redirect()->route('eventos.index');
         }
         $evento->update(["estado_moderacion" => "rechazado"]);
-        return redirect()->route('comunidades.show',$evento->comunidad->user_id);
+        return redirect()->route('eventos.index');
     }
 }
